[TASK] Add ArrayUtility and MathUtility
This is basically a forward port of the TYPO3 CMS extension
EXT:nxcrudextbase. There is just some tuning to make it work
with TYPO3 Flow but neither changes in internal behavior nor
renaming of objects and properties.

The ArrayUtility is really part of EXT:nxcrudextbase. The
MathUtility was part of TYPO3 CMS.

Change-Id: Idfdeaee63213ea3fcb240d4fa1e1fc8b24ad2264
Related: INFRA-10